Add files via upload
This is the V3 of the EMS. I've finished nearly all the frontend. I've committed out some areas where I set the ground work but wasn't sure what was neccessary for the backend for it to work. The pages that need attention are the Account Info Pages for the patient, the Patient Directory Page, the Notes/Document page, and the Patient Reservation page. 

Each page should be setup in the neccessary fashion to send information to the server and utlize it. The reservations page for the customer is using google calendars, as we will give certain staff users access to this calendar to setup doctor avilaibilities and accept patient reservation requests. In addition payments that are made are simply just emailed to the patient and an outside banking system will be used for monetary transactions.